Initial Commit
Create an application that searches for the top images of the week from
the Imgur gallery and displays them in a list.
The app needs to provide the user with the ability to input text as part of
the search query before displaying relevant results.
Each cell needs to display the following for each search result:
• title
• date of the post in local time (DD/MM/YY hh:mm am/pm)
• number of additional images in the post (if there are multiple)
• image
A search result should display the first image from its list if it contains
multiple images.